requirements.txt
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.4.0
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=0.24.0
tqdm>=4.60.0
networkx>=2.5.0
psycopg2-binary>=2.9.0
torch-geometric>=2.0.0
transformers>=4.5.0
seaborn>=0.11.0
scipy>=1.7.0
shap>=0.40.0
captum>=0.5.0
ipywidgets>=7.6.0
jupyter>=1.0.0
jupyterlab>=3.0.0
plotly>=5.0.0
tensorboard>=2.5.0
pytorch-lightning>=1.4.0
dask>=2021.6.0
boto3>=1.18.0
nltk>=3.6.0
spacy>=3.0.0 